+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* `tbl_survival()`: the `groupname` column name has been changed to `label_level` (#133)
+
 # gtsummary 0.1.0 (2019-05-04)
 
 * First public release